Crate pallet_contracts_primitives

A crate that hosts a common definitions that are relevant for the pallet-contracts.

Enums

ContractAccessError
The possible errors that can happen querying the storage of a contract.

RentProjection

Type Definitions

GetStorageResult
A result type of a get storage call.

RentProjectionResult
A result type of a rent_projection call.