Crate pallet_contracts_primitives

A crate that hosts a common definitions that are relevant for the pallet-contracts.

Structs

CodeUploadReturnValue

The result of successfully uploading a contract.

ContractResult

Result type of a bare_call or bare_instantiate call as well as ContractsApi::call and ContractsApi::instantiate.

ExecReturnValue

Output of a contract call or instantiation which ran to completion.

InstantiateReturnValue

The result of a successful contract instantiation.

ReturnFlags

Flags used by a contract to customize exit behaviour.

Enums

Code

Reference to an existing code hash or a new wasm module.

ContractAccessError

The possible errors that can happen querying the storage of a contract.

StorageDeposit

The amount of balance that was either charged or refunded in order to pay for storage.

Type Aliases

CodeUploadResult

Result type of a bare_code_upload call.

ContractExecResult

Result type of a bare_call call as well as ContractsApi::call.

ContractInstantiateResult

Result type of a bare_instantiate call as well as ContractsApi::instantiate.

GetStorageResult

Result type of a get_storage call.