Crate pallet_contracts_primitives

A crate that hosts a common definitions that are relevant for the pallet-contracts.

Structs

ContractExecResult
Result type of a bare_call call.

ExecError
Error returned by contract exection.

ExecReturnValue
Output of a contract call or instantiation which ran to completion.

ReturnFlags
Flags used by a contract to customize exit behaviour.

Enums

ContractAccessError
The possible errors that can happen querying the storage of a contract.

ErrorOrigin
Origin of the error.

RentProjection

Type Definitions

ExecResult
The result that is returned from contract execution. It either contains the output buffer or an error describing the reason for failure.

GetStorageResult
Result type of a get_storage call.

RentProjectionResult
Result type of a rent_projection call.